Replace CacheTagInvalidator with CacheTagManager
